Rename G2Names.hpp to IconNames.hpp
