+++ src/openrct2-ui/scripting/ScWidget.hpp
@@ -21,7 +21,7 @@
 #    include <openrct2/Context.h>
 #    include <openrct2/common.h>
 #    include <openrct2/scripting/Duktape.hpp>
-#    include <openrct2/scripting/G2Names.hpp>
+#    include <openrct2/scripting/IconNames.hpp>
 #    include <openrct2/scripting/ScriptEngine.h>
 
 namespace OpenRCT2::Scripting
