+++ src/openrct2/scripting/IconNames.hpp
