+++ src/openrct2/scripting/bindings/game/ScContext.hpp
@@ -18,8 +18,8 @@
 #    include "../../../object/ObjectManager.h"
 #    include "../../../scenario/Scenario.h"
 #    include "../../Duktape.hpp"
-#    include "../../G2Names.hpp"
 #    include "../../HookEngine.h"
+#    include "../../IconNames.hpp"
 #    include "../../ScriptEngine.h"
 #    include "../game/ScConfiguration.hpp"
 #    include "../game/ScDisposable.hpp"
